@@ -7,9 +7,12 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using StarWars.Web.Brokers;
 using StarWars.Web.Services;
+using StarWars.Web.Models;
 using System.Net.Http.Headers;
-using StarWars.Web.Context;
 using Microsoft.EntityFrameworkCore;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.HttpsPolicy;
+using Microsoft.Extensions.Hosting;
 
 namespace StarWars.Web
 {
@@ -25,13 +28,6 @@ namespace StarWars.Web
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            services.Configure<CookiePolicyOptions>(options =>
-            {
-                // This lambda determines whether user consent for non-essential cookies is needed for a given request.
-                options.CheckConsentNeeded = context => true;
-                options.MinimumSameSitePolicy = SameSiteMode.None;
-            });
-
             services.AddHttpClient("swapiClient", client =>
             {
                 client.BaseAddress = new Uri("https://swapi.co/api/");
@@ -40,19 +36,11 @@ namespace StarWars.Web
 
             services.AddTransient<IStarWarsApiClient, StarWarsApiClient>();
 
-            services.AddSingleton<ILoggingBroker, LoggingBroker>();
-
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
-            var connection = Configuration.GetConnectionString("DefaultConnection");
-            services.AddDbContext<StorageContext>(options =>
-            {
-                options.UseSqlServer(connection);
-                options.UseQueryTrackingBehavior(QueryTrackingBehavior.NoTracking);
-            });
+            services.AddMvc();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
@@ -67,13 +55,12 @@ namespace StarWars.Web
 
             app.UseHttpsRedirection();
             app.UseStaticFiles();
-            app.UseCookiePolicy();
-
-            app.UseMvc(routes =>
+            app.UseRouting();
+            app.UseEndpoints(endpoints =>
             {
-                routes.MapRoute(
+                endpoints.MapControllerRoute(
                     name: "default",
-                    template: "{controller=Home}/{action=Index}/{id?}");
+                    pattern: "{controller=Home}/{action=Index}/{id?}");
             });
         }
     }
