@@ -2,26 +2,22 @@ using System;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
 using StarWars.Web.Models;
-using StarWars.Web.Context;
 namespace StarWars.Web.Brokers
 {
-    public class StorageBroker : IStorageBroker<People>
+    public class StorageBroker : DbContext, IStorageBroker<People>
     {
-        private readonly StorageContext strorageContext;
-        public StorageBroker(StorageContext storageConext)
-            => this.strorageContext = storageConext;
-
+        // private DbSet<People> Peoples { get; set; }
         public async Task AddEntity(People people)
         {
-            try
-            {
-                await this.strorageContext.Peoples.AddAsync(people);
-                await this.strorageContext.SaveChangesAsync();
-            }
-            catch (DbUpdateException dbUpdateException)
-            {
-                throw dbUpdateException;
-            }
+            // try
+            // {
+            //     await this.Peoples.AddAsnyc(people);
+            //     await this.SaveChangesAsync();
+            // }
+            // catch (DbUpdateException dbUpdateException)
+            // {
+            //     throw dbUpdateException;
+            // }
         }
     }
 }
