@@ -6,9 +6,6 @@ namespace StarWars.Web.Services
     public interface IStarWarsApiClient
     {
         // api returns data in pages, passing page number as parameter
-        Task<PeopleRootObject> GetAsync(string queryString);
-
-        // Add extra methods for neew entities
-        Task AddPeopleAsnyc(People people);
+        Task<PeopleRootObject> GetAsync(string entity, string pageUrl);
     }
 }
